Development (#53)
* closes #31


* closes #19

* closes #30

* stackexchange exceptional removed, elmah added

* elmah

* closes #29

* close #25

* close

* #33

* #33

* issue

* #33 first test


* #34

* #34

* fix error handling

* fix build

* add jasmine locally for build

* add packages for build

* try to fix build

* code coverage #34

* fix build

* closes #34

* shields

* code cov

* fix coverage

* fix code coverage

* fix coverage

* splitting components + accountservice cleanup (1)

* #27

* #27

* #23

* #23 add more

* #27

* closes #36

* #27 #21

* #27 order docdb logic

* fix tests #27

* tests

* codecov

* fix backend build

* fix codecov?

* #21

* test

* various fixes

* fixes menu

* fix observable package

* resolve implicit any type

* resolve implicit any type 2

+++ backend/WebApi/Lunchorder.Domain/Dtos/LastOrder.cs
@@ -1,7 +1,5 @@
 using System;
 using System.Collections.Generic;
-using Newtonsoft.Json.Serialization;
-using Newtonsoft.Json;
 
 namespace Lunchorder.Domain.Dtos
 {
@@ -36,6 +34,6 @@ namespace Lunchorder.Domain.Dtos
         /// <summary>
         /// The total price of all entries of this order including rules
         /// </summary>
-        public double FinalPrice { get; set; }
+        public decimal FinalPrice { get; set; }
     }
 }
